+++ src/xyz/oihana/schema/constants/traits/statistics/ObservationSeriesTrait.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\statistics;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\statistics\ObservationSeries} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\statistics
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.4.0
+ */
+trait ObservationSeriesTrait
+{
+    const string VALUES = 'values' ;
+}