+++ src/xyz/oihana/schema/constants/traits/statistics/StatisticsRecordTrait.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\statistics;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\statistics\Statistics} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\statistics
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.4.0
+ */
+trait StatisticsRecordTrait
+{
+    const string ABOUT              = 'about'             ;
+    const string ASSIGNED_COMPANY   = 'assignedCompany'   ;
+    const string DIRECTION          = 'direction'         ;
+    const string OBSERVATION_PERIOD = 'observationPeriod' ;
+    const string YEAR               = 'year'              ;
+}